+++ showData.h
@@ -14,15 +14,15 @@ protected:
     int showRating;
     std::string showGenre;
 public:
-    Show(std::string showName = "null", int episodesWatched = 0, int episodesTotal = 0, int showRating = 0, std::string showGenre = "null");
+    Show(std::string showName = "null", int episodesWatched = 0, int episodesTotal = 0, double showRating = 0, std::string showGenre = "null");
     void setTitle(std::string newTitle);
     std::string getTitle();
     void setWatched(int newWatched);
     int getWatched();
     void setEpisodes(int newEpisodes);
     int getEpisodes();
-    void setRating(int newRating);
-    int getRating();
+    void setRating(double newRating);
+    double getRating();
     void setGenre(std::string newGenre);
     std::string getGenre();
